main mr전 마지막 점검

See merge request cloud_track/class_04/web_project3/team06/hr-handlers-be!38

+++ src/main/java/com/hr_handlers/admin/dto/employee/EmpRegisterDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDate;
 public class EmpRegisterDto {
     private String empNo;
     private String password;
-    private Long leaveBalance;
+    private double leaveBalance;
     private Role role;
     private ContractType contractType;
     private LocalDate joinDate;
